Moved profiling.py to shared_utils/.
